In _get_vfile(), handle the case when a referenced object has been punched
After a crash or power failure, we may still have an entry for an object
in the index server, that had just been punched from the volume prior to
the crash.

When that happens, do not raise an error, remove the entry from the
index server. This is already done in a couple of other places, but was
missing here.

Change-Id: Icbcfc7148fdaeeb21df4b4173c2d99339639a550