Erase object header if we fail to register it to the index-server
When we commit a file, we write an object header, the swift metadata,
and register the object in the index server. If that last steps fails, an
exception is raised and the commit fails.

Without this patch, the object header remains in the volume. If a power loss
or crash occurs before that header is overwritten by a new object, the
volume-check script will pick up the object when the machine is rebooted.

With the patch, the header is overwritten with null bytes. (Punching a hole
is slower and will create an extent).

While doing this, add a constant for the maximum serialized header length

Change-Id: I618f061511b65cd9b820f8e995e6605562a82acd