Import directly from the inside of the ring module
An odd thing happened: when my code did only
"from swift.common.ring import Ring", and nothing else,
the interpreter looped on the CPU.

The import from the top of the ring module is our standard
protocol for Ring. It causes no problem in places like
container updater or list_endpoints.py. It is a big
mystery why it causes Python 3.7.3 to loop, and only
in specific circumstances.

But we noticed that the recursive import is unnecesasry
in this case, so an obviously harmless fix exists.

Change-Id: I7373bbb0a50d090d6becf141e5832f8ae98381a4