container-sharding: Stable overlap order
When you have overlapping active shard ranges updates will get sent to
"the first" database; but when the proxy queries shards for listings
they get stitched together end-to-end with markers.

This means mostly the second shard range is ignored.  But since the
order of shard ranges is not stable (it seems to depend on the database
indexes; which can change when rows are added or removed) you could send
updates to "the wrong" shard.

Using a stable order leads to more correct and robust behavior under
failure; and is also better for cognitive overhead.

Cherry-Picked-From: Ia9d29822bf07757fc1cf58ded90b49f12b7b2c24
Change-Id: Ibf304bb64f005a86cb4d08312c5d7ad4a2488719