Add librsvg2-* to bindep.txt
Add the librsvg2-* package to bindep.txt. This package has been
required to build docs since the Related-Change. The package is added
to a new 'doc' profile in bindep.txt; if missing, the package will be
reported by:

  bindep doc

The 'doc' profile is added to the tox 'bindep' env command so that, if
missing, the package will always be reported by:

  tox -e bindep

Change-Id: I6c60c31ca8002133ab77f05f59359b25315b299c
Related-Change: I26cefda80d3234df68d7152b404e0a71da74ab90
Closes-Bug: 1903038