Add a read-only role to keystoneauth
An idea was floated recently of a read-only role that can be used for
cluster-wide audits, and is otherwise safe. It was also included into
the "Consistent and Secure Default Policies" effort in OpenStack,
where it implements "reader" personas in system, domain, and project
scopes. This patch implements it for system scope, where it's most
useful for operators.

Change-Id: I5f5fff2e61a3e5fb4f4464262a8ea558a6e7d7ef